move drag-drop.ts into services folder
