+++ packages/client/hmi-client/src/page/data-explorer/components/asset-card.vue
@@ -112,7 +112,7 @@ import { Model } from '@/types/Model';
 import { Dataset } from '@/types/Dataset';
 import { ResourceType } from '@/types/common';
 import * as textUtil from '@/utils/text';
-import { useDragEvent } from '@/utils/drag-drop';
+import { useDragEvent } from '@/services/drag-drop';
 
 // This type is for easy frontend integration with the rest of the extraction types (just for use here)
 type UrlExtraction = {
