+++ packages/client/hmi-client/src/services/drag-drop.ts
